requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pillow==10.4.0
pygame==2.6.1
pydantic==2.9.2
watchdog==4.0.2
pyyaml==6.0.2